Start conversation binds to the service.
We need to figure out a way for the service to handle the incoming messages at any time. For example, it may receive a new chat message while the user is browsing the "commute" screen. The service should perform the correct operation at message arrivals and optionally notify the user.